+++ anlffr/bootstrap.py
@@ -1,405 +1,246 @@
-# -*- coding: utf-8 -*-
-"""
-Bootstrap helper functions for FFR data
-"""
 import numpy as np
+from anlffr import spectral
+from .utils import logger, verbose
 import time
-import multiprocessing
-import errno
-from .utils import logger
-from .utils import verbose as verbose_decorator
-import platform
 
-
-@verbose_decorator
-def bootfunc(inputFunction, x, params, verbose=True):
-    """Performs bootstrapping over trials for spectral functions.
-
-    Inputs
-    ------
-    inputFunction - Function to run on x (the actual function, not the string
-    name of the function). Function must take numpy 3D array of form chan x rep
-    x time as the first argument and a paramater dictionary as the second
-    argument. Function must return a dictionary.
-
-    x - a 3D numpy array or list/tuple of 3D numpy arrays (note: must be chan x
-    trial x time). If list/tuple with N elements, an equal number of trials M
-    will be selected from each pool at random without replacement, based on the
-    minimum number of trials available in eahc pool. Those will then be sampled
-    with replacement to obtain bootstrapped mean/variance, i.e., results will
-    be computed by combining M from each pool per bootstrap repetition. An
-    example use case is to sample an equal number of trials from positive and
-    negative polarity FFR trials.
-
-    params - dictionary of parameters. The following fields are required for
-    boostrapping, but others may be required based on the specifics of
-    inputFunction:
-
-          params['nDraws'] - number of draws for bootstrapping
-
-          params['threads'] - number of threads to spawn; only really useful if
-          a multi-core CPU is available. When a multi-core CPU is available,
-          computations should be speeded by a factor of approximately
-          params['threads'].
-
-    Returns
-    -------
-    out - Dictionary of dictionaries. Level 1 keys corrspond to
-    function outputs, level 2 keys correspond to individual draw results, mean
-    of draw results, and variance of draw results for each key in level 1.
-
-    Examples
-    --------
-    Just a simple illustration of usage.
-
-    >>> # Package imports and initialize parameters
-    >>> from anlffr import bootstrap, spectral
-    >>> params = spectral.generate_params(...,
-                                         threads = 4,
-                                         nDraws = 100)
-
-    >>> positivePolarityData = io.loadmat(...)['data']
-    >>> negativePolarityData = io.loadmat(...)['data']
-
-    >>> # create a list of data
-    >>> dataList = [positivePolaritydata, negativePolaritydata]
-
-    >>> # run the boostrap function
-    >>> results = bootstrap.bootfunc(spectral._mtcpca_complete, dataList,
-                                     params)
-
-    Notes
-    -------
-    Utilizes multiple cores (threads) for a speed increase.
-    Designed to work with the spectral analysis functions provided in ANLffr,
-    but in theory should work with any function that takes 3D numpy arrays
-    and a suitable params structure as inputs.
-
-    Bootstrapping is not implemented in a particularly clever or elegant way -
-    it's an "emarassingly parallel" problem, and thus was coded in the
-    bare-minimum fashion so that boostrap computations can proceed in parallel
-    and change the amount of time you wait for a final result by 1/nThreads.
-    Note that this speedup will only take place on Linux/Mac (POSIX in
-    general?) systems - not on Windows machines. The current implementation
-    does not play well with process forks on Windows, and as such, any attempt
-    to run things in multithreaded mode will fall back to a single thread.
-
-    The data array is copied in memory for each core that is utilized, and will
-    thus eat quite a bit of RAM when more than 1-2 threads are spawned. RAM is
-    cheap, so this shouldn't be too much of a concern.
-
-    Last updated: 2016-04-03
-    Auditory Neuroscience Laboratory, Boston University
-    Contact: lennyv@bu.edu
-    """
-
-    _validate_bootstrap_params(params)
-
-    sanitizedData, nTrialsUsed = _equate_data(x, params)
-
+@verbose
+def bootfunc(inputFunction, x1, params, verbose=True):
+    '''
+    computes the bootstrapped mean and variance of x1 as processed by
+    inputFunction. inputFunction is expected to be one of the functions from
+    anlffr.spectral.
+    '''
     startTime = time.time()
-    theQueue = multiprocessing.Queue()
-    results = {}
+    params = _fix_params(params)
+    try:
+        from joblib import Parallel, delayed
+        nJobs = int(params['threads'])
+    except ImportError:
+        logger.warning('joblib not installed; cannot run in parallel.')
+    except KeyError:
+        nJobs = 1
+
+    nDraws = int(params['nDraws'])
+
+    x1, n = _equate_within_pool(x1)
+    if nJobs == 1:
+        results = []
+        for i in range(nDraws):
+            results.append(_run_bootfunc(inputFunction, x1, params))
+    else:
+        P = Parallel(n_jobs=nJobs)
+        results = P(delayed(_run_bootfunc)(inputFunction, x1, params) 
+                    for i in range(nDraws))
+
+    concatenated = _dict_concatenate(results)
+
     output = {}
-
-    # split the loads roughly equally:
-    drawSplit = _compute_thread_split(params)
-
-    # set up the processes
-    processList = []
-    for proc in range(len(drawSplit)):
-        if 'debugMode' in params and params['debugMode']:
-            logger.warn('Warning: setting fixed random seeds!')
-            randomState = np.random.RandomState(proc)
-        else:
-            randomState = np.random.RandomState(None)
-
-        if platform.system() != 'Windows':
-            processList.append(
-                multiprocessing.Process(target=_multiprocess_wrapper,
-                                        args=(inputFunction,
-                                              sanitizedData,
-                                              params,
-                                              drawSplit[proc],
-                                              theQueue,
-                                              randomState)))
-
-            processList[proc].start()
-        else:
-            _multiprocess_wrapper(inputFunction,
-                                  sanitizedData,
-                                  params,
-                                  drawSplit[proc],
-                                  theQueue,
-                                  randomState)
-
-    numRetrieved = 0
-    trialsUsed = []
-    frequencyVector = None
-
-    status = 'Retrieved {:d}/{:d} '
-    while numRetrieved < params['nDraws']:
-        try:
-            retrievedData = theQueue.get(True)
-
-            usefulKeys = retrievedData[0].keys()
-
-            # only need to retrieve frequency vector once
-            # otherwise remove it from key values to store
-            if 'f' in usefulKeys:
-                if frequencyVector is None: 
-                    frequencyVector = retrievedData[0]['f']
-                else:
-                    if not all(frequencyVector == retrievedData[0]['f']):
-                        logger.error('frequency vector mismatch across draws.')
-
-                usefulKeys.remove('f')
-
-            # now run through the other keys and store the results
-            for k in usefulKeys:
-                # set up the dictionary fields with the first retrieved piece
-                if 0 == numRetrieved:
-                    results[k] = dict(runningSum=0, runningSS=0, indivDraw=[])
-
-                if params['returnIndividualBootstrapResults']:
-                    results[k]['indivDraw'].append(retrievedData[0][k])
-
-                results[k]['runningSum'] += retrievedData[0][k]
-                results[k]['runningSS'] += retrievedData[0][k]**2.0
-
-            trialsUsed.append(retrievedData[1])
-            
-            numRetrieved = numRetrieved + 1
-            logger.info(status.format(numRetrieved, params['nDraws']))
-
-        # the following should be OK, as per:
-        # http://stackoverflow.com/questions/
-        #        4952247/interrupted-system-call-with-processing-queue
-        except IOError as e:
-            if e.errno == errno.EINTR:
-                continue
-            else:
-                raise
-
-    # ensure the queue is empty:
-    # Note: this results in a bug on osx because qsize not implemented
-    # if int(theQueue.qsize()) > 0:
-    #     logger.error('Internal error: retrieved all data, but Queue is ' +
-    #                  'nonempty: {}'.format(int(theQueue.qsize())))
+    usefulKeys = list(concatenated.keys())
+    print(usefulKeys)
+    output['f'] = concatenated['f']
+    usefulKeys.remove('f')
 
     for k in usefulKeys:
         output[k] = {}
-        output[k]['nDraws'] = int(params['nDraws'])
-        output[k]['nPerDraw'] = nTrialsUsed 
+        output[k]['bootMean'] = np.mean(concatenated[k], axis=0)
+        output[k]['bootVariance'] = np.var(concatenated[k], axis=0)
+        output[k]['nDraws'] = nDraws
+        output[k]['nPerDraw'] = n
         if params['returnIndividualBootstrapResults']:
-            output[k]['indivDraw'] = np.array(results[k]['indivDraw'])
-        output[k]['bootMean'] = results[k]['runningSum'] / params['nDraws']
-        output[k]['bootVariance'] = _compute_variance(output[k]['bootMean'],
-                                                      results[k]['runningSS'],
-                                                      params['nDraws'])
-
-    if frequencyVector is not None:
-        output['f'] = frequencyVector
+            output[k]['indivDraw'] = concatenated[k]
 
     logger.info('\nCompleted in: {} s'.format(time.time() - startTime))
-
     return output
 
 
-@verbose_decorator
-def _multiprocess_wrapper(inputFunction, inputData, params, nDraws, 
-                          resultsQueue, randState, verbose=True):
-    """
-    internal function. places results from spectral functions in queue.
-    """
+@verbose
+def permutation_distributions(inputFunction, x1, x2, params, verbose=True):
 
-    for _ in range(nDraws):
-
-        theseData, theseParams, trialsUsed = _select_trials_with_replacement(
-            inputData, params, randState)
-
-        out = (inputFunction(theseData, theseParams, verbose=False,
-                             bootstrapMode=True),
-               trialsUsed)
-
-        resultsQueue.put(out)  # block by default...
-
-
-@verbose_decorator
-def _compute_variance(dataMean, dataSumOfSquares, n, verbose=None):
-    """
-     internal function. computes variance from running SS, mean, and n.
-    """
-    return (dataSumOfSquares - ((dataMean * n) ** 2) / n) / (n - 1)
-
-
-@verbose_decorator
-def _select_trials_with_replacement(inputData,
-                                    params,
-                                    randomState=None,
-                                    verbose=True):
-    """
-    internal function. creates a new data array from a series of randomly
-    sampled old ones. random sample is with replacement.
-    """
-
-    if not isinstance(inputData, list):
-        logger.error('Internal error: inputData should be a list of arrays.')
-
-    # should make a copy of params to avoid original being modified
-    modifiedParams = dict(params)
-
-    if randomState is None:
-        randomState = np.random.RandomState()
-
-    numPools = len(inputData)
-    useTrialsPerPool = inputData[0].shape[1]
-
-    tempData = []
-    pickTrials = []
-    tempSelected = []
-
-    for pool in range(numPools):
-
-        randTrials = randomState.randint(0, inputData[pool].shape[1],
-                                         useTrialsPerPool)
-
-        tempData.append(inputData[pool][:, randTrials, :])
-        pickTrials.append(randTrials)
-
-        # because things will be concatenated, add useTrialsPerPool*pool
-        # to each value in randTrials. Functions in spectral.py will only
-        # use even-labeled trials When computing the noise floor.
-        tempSelected.append(randTrials + useTrialsPerPool*pool)
-
-    useData = np.concatenate(tuple(tempData), axis=1)
-
-    modifiedParams['bootstrapTrialsSelected'] = np.concatenate(
-        tuple(tempSelected))
-
-    return (useData, modifiedParams, pickTrials)
-
-
-@verbose_decorator
-def _compute_thread_split(params, verbose=None):
-    """
-    internal function. computes how many draws each process will take on.
-    """
-    drawSplit = []
-    if platform.system() != 'Windows':
-        distribute = int(params['nDraws']) / int(params['threads'])
-        leftover = int(params['nDraws']) % int(params['threads'])
-
-        if distribute != 0:
-            for _ in range(params['threads']):
-                drawSplit.append(int(distribute))
-
-            for r in range(leftover):
-                drawSplit[r] = int(drawSplit[r] + 1)
-        else:
-            for _ in range(int(params['nDraws'])):
-                drawSplit.append(1)
-    else:
-        drawSplit.append(params['nDraws'])
-
-    return drawSplit
-
-
-def _equate_data(inputData, params):
-    """
-    shuffles the trials and ensures that each pool to draw from has the same
-    number of trials by selecting the minimum number in common across all pools
-    """
-
-    # allows you to specify a list of data from each polarity (or other things
-    # you want to combine across) so that the number of trials going into a
-    # computation from different sources can be fixed
+    '''
+    Returns x1-x2 when each is computed using inputFunction, and also the
+    distribution of differences when the labels are shuffled at random.
+    inputFunction is expected to be one of the functions from anlffr.spectral.
+    '''
+    startTime = time.time()
+    params = _fix_params(params)
     try:
-        if isinstance(inputData, list) or isinstance(inputData, tuple):
-            for x in inputData:
-                if not isinstance(x, np.ndarray) and x.ndim != 3:
-                    raise TypeError
+        from joblib import Parallel, delayed
+        nJobs = int(params['threads'])
+    except ImportError:
+        logger.warning('joblib not installed; cannot run in parallel.')
+    except KeyError:
+        nJobs = 1
 
-        elif isinstance(inputData, np.ndarray):
-            inputData = [np.array(inputData)]
+    nDraws = int(params['nDraws'])
 
-    except TypeError:
-        logger.error('Data should be 3D numpy array or list/tuple ' +
-                     'of 3D numpy arrays')
+    x1, n1 = _equate_within_pool(x1)
+    x2, n2 = _equate_within_pool(x2)
+    
+    x1Res = inputFunction(np.concatenate(x1, axis=1), params)
+    x2Res = inputFunction(np.concatenate(x2, axis=1), params)
+    difference = _dict_diff(x1Res, x2Res)
 
-    poolSizes = []
+    if nJobs == 1:
+        results = []
+        for i in range(nDraws):
+            results.append(get_null_difference(inputFunction, x1, x2, params))
 
-    for x in inputData:
-        poolSizes.append(x.shape[1])
-
-    minimumAcrossPools = min(poolSizes)
-
-    if minimumAcrossPools % 2 == 1:
-        minimumAcrossPools -= 1
-
-    # make sure the user knows this is what is going on
-    warnStr1 = ('Selecting {} per pool '.format(minimumAcrossPools) +
-                'WITHOUT replacement from original data supplied')
-
-    warnStr2 = ('Will select from this subset WITH replacement to ' +
-                'compute bootstrap mean/variances')
-
-    logger.info(warnStr1 + '\n' + warnStr2)
-
-    validatedData = []
-
-    if 'debugMode' in params and params['debugMode']:
-        randState = np.random.RandomState(31415)
     else:
-        randState = np.random.RandomState(None)
+        P = Parallel(n_jobs=nJobs)
+        results = P(delayed(_get_null_difference)(inputFunction, x1, x2, params)
+                    for i in range(nDraws))
 
-    # shuffle the trials within each pool and equate the number of trials:
+    nullDifferenceDistribution = _dict_concatenate(results)
+
+    return difference, nullDifferenceDistribution
+
+
+@verbose
+def _dict_diff(x1Res, x2Res, verbose=True):
+    '''
+    for each key in a dictionary, subtract x2 from x1. both assumed to be
+    arrays.
+    '''
+    difference = {}
+    usefulKeys = list(x1Res.keys())
+    usefulKeys.remove('f')
+
+    for k in usefulKeys:
+        difference[k] = x1Res[k] - x2Res[k]
+
+    difference['f'] = x1Res['f']
+
+    return difference
+
+
+@verbose
+def _dict_concatenate(resList, verbose=True):
+    '''
+    Concatenates arrays in a list of dictionaries, as would be returned when
+    running anlffr.spectrum functions using joblib parallel instance.
+    '''
+    concatenated = {}
+    usefulKeys = list(resList[0].keys())
+    concatenated['f'] = resList[0]['f']
+    usefulKeys.remove('f')
+    
+    for k in usefulKeys:
+        toConcatenate = []
+        for y in range(len(resList)):
+            toConcatenate.append(resList[y][k])
+        concatenated[k] = np.vstack(toConcatenate)
+
+    return concatenated
+
+
+@verbose
+def _run_bootfunc(inputFunction, x1, params, verbose=True):
+    '''
+    the function actually being fed to joblib parallel for bootstrap
+    computation of mean and variance
+    '''
+    x1s = _sample_with_replacement(x1)
+    x1sRes = inputFunction(np.concatenate(x1s, axis=1), params)
+
+    return x1sRes 
+
+
+@verbose
+def _get_null_difference(inputFunction, x1, x2, params, verbose=True):
+    '''
+    the function actually being fed to joblib parallel for permutation testing 
+    '''
+    x1s, x2s = _label_shuffler(x1, x2)
+    x1sRes = inputFunction(np.concatenate(x1s, axis=1), params)
+    x2sRes = inputFunction(np.concatenate(x2s, axis=1), params)
+
+    nullDiff = _dict_diff(x1sRes, x2sRes) 
+
+    return nullDiff
+
+
+@verbose
+def _label_shuffler(x1, x2, verbose=True):
+    '''
+    randomly reassigns the trials belonging to x1 and x2, and returns two
+    arrays the same size as the original input
+    '''
+    r = np.random.RandomState(None)
+    x1s = []
+    x2s = []
+    for y in range(len(x1)):
+        x1s.append(np.empty(x1[y].shape))
+        x2s.append(np.empty(x2[y].shape))
+
+    for y in range(len(x1)):
+        temp = np.concatenate([x1[y], x2[y]], 1)
+        tempOrder = r.permutation(temp.shape[1])
+        temp = temp[:, tempOrder, :]
+        x1s[y] = temp[:, 0:x1[y].shape[1], :]
+        x2s[y] = temp[:, x1[y].shape[1]:, :]
+
+    return x1s, x2s
+
+
+@verbose
+def _equate_within_pool(inputData, verbose=True):
+    '''
+    Sets the number of trials per list element to be equal For example: useful
+    when computing EFRs, and need an equal number of +/- polarity trials in
+    computation
+    '''
+    r = np.random.RandomState(None)
+    errorStr = 'list/tuple of 3D arrays'
+    if isinstance(inputData, np.ndarray):
+        if len(inputData.shape) == 3:
+            inputData = [inputData]
+        else:
+            raise ValueError(errorStr)
+
+    minAvail = np.Inf
     for x in inputData:
-        # select the same number of trials from each pool without replacement
-        randomOrder = randState.permutation(x.shape[1])
-        randomOrder = randomOrder[0:minimumAcrossPools]
-        validatedData.append(x[:, randomOrder, :])
+        if not isinstance(x, np.ndarray) or x.ndim != 3:
+            raise ValueError(errorStr)
+        if x.shape[1] < minAvail:
+            minAvail = x.shape[1]
+    n = 0
+    for x in range(len(inputData)):
+        keepTrials = r.permutation(inputData[x].shape[1])[0:minAvail]
+        n += len(keepTrials)
+        inputData[x] = inputData[x][:, keepTrials, :]
 
-    return validatedData, minimumAcrossPools*len(inputData)
+    return inputData, n
 
 
-@verbose_decorator
-def _validate_bootstrap_params(params, verbose=True):
-    """
-    internal function. checks parameters required for bootfunc and throws an
-    error if conditions are violated
-    """
+@verbose
+def _sample_with_replacement(inputData, verbose=True):
+    '''
+    generates a dataset the same size as the original that was constructed by
+    sampling the original with replacement
+    '''
+    r = np.random.RandomState(None)
+    errorStr = 'Expecting a list/tuple of 3D arrays'
+    if isinstance(inputData, np.ndarray):
+        if len(inputData.shape) == 3:
+            inputData = [inputData]
+        else:
+            raise ValueError(errorStr)
 
-    if platform.architecture()[0] != '64bit':
-        logger.warning('Python is running in 32 bit mode. ' +
-                       'You may encounter out-of-memory issues.')
+    resampled = list(inputData)
 
-    if 'nDraws' not in params:
-        logger.error('params[''nDraw''] must be specified')
-        
-    if params['nDraws'] <= 0:
-        logger.error('params[''nDraws''] must be positive')
+    for x in range(len(inputData)):
+        if not isinstance(inputData[x], np.ndarray) or inputData[x].ndim != 3:
+            raise ValueError(errorStr)
+        tr = r.randint(inputData[x].shape[1], size=inputData[x].shape[1])
+        resampled[x] = inputData[x][:, tr, :]
 
-    if params['nDraws'] != int(params['nDraws']):
-        logger.error('params[''nDraws''] must be an integer')
+    return resampled
 
-    if platform.system() == 'Windows':
-        logger.warn('Windows system detected...' +
-                    'will only use one execution thread.')
-        checkThreads = False
-    else:
-        checkThreads = True
+@verbose
+def _fix_params(params):
+    fixedParams = dict(params)
+    if ('bootstrapMode' not in params.keys() or
+        fixedParams['bootstrapMode'] == False):
+        fixedParams['bootstrapMode'] = True
 
-    if checkThreads and 'threads' in params:
-        logger.info('Attempting to use ' +
-                    '{} threads...'.format(params['threads']))
-        numCpu = multiprocessing.cpu_count()
-
-        if 0 > params['threads']:
-            logger.error('params[''threads''] should be > 0')
-
-        if params['threads'] > numCpu:
-            logger.warn(
-                'params[''threads''] should optimally be <= {}'.format(numCpu))
-
-    return
+    return fixedParams
